BACKEND WORK
I hooked up the compojure.api with all the relevant libraries and made a
functioning post-fetching route. There's also some test data in a
comment in /src/clj/blog/posts/db.clj to insert to db if you need to
test this commit :D

This thing needs to be componentized. Then I have to hook the
reloaded.repl to this thing