BACKEND
The way db spec is injected to the route's scope finally pleases me. You
need to have a :sys sys in the route's (or parent context's) definition,
and after putting that symbol you used after :sys into the
blog.util/destructure-db macro, you have a hold of the db spec.