BOTH
There is support and UI for posting and deleting comments, and edit
buttons that don't do a thing.

Frontend system is a bit revamped. System supports now defining routes
with secretary and thus the use of :a's instead of :buttons is now
encouraged. Setting session's :current-main to a parameterless reagent
component renders that as the main component. Session's :current-sidebar
is the express route to the sidebar.