Landing page implemented on the frontend, #6 almost done
The only reasons I'm not able to close it: frontend should be made to
query if there is already a landing page tagged when saving a new
post. After implementing that, one might wonder if this subsystem
should be refactored to make it relationally provable that there is
only one landing page.